move constants to style
