+++ screens/BackgroundSelectorScreen.tsx
@@ -2,7 +2,7 @@ import React, { FC, useContext, useEffect } from "react";
 import { Platform, View, Text, TouchableHighlight } from "react-native";
 import * as ImagePicker from "expo-image-picker";
 import { BackgroundImageContext } from "../contexts/BackgroundImageContext";
-import { styles } from "../constants/Styles";
+import { styles } from "../style/Styles";
 import { ScreenOrientationContext } from "../contexts/ScreenOrientationContext";
 import { Orientation } from "expo-screen-orientation";
