+++ style/SensorsStyles.ts
